refactor(merge): split pure text/duration builders into MergeDialog.DialogText.cs
