+++ scripts/ers-els-audit-allowlist.txt
@@ -167,6 +167,13 @@ Source/Parsek/CrewReservationManager.cs
 # filters it out by definition — raw read is the only way to find it.
 Source/Parsek/MergeDialog.cs
 
+# MergeDialog.DialogText.cs is the partial-class split holding the pure
+# text/duration builders. FindReFlyRecording (moved verbatim from
+# MergeDialog.cs) reads RecordingStore.CommittedRecordings directly to
+# locate the provisional re-fly recording by id when it is not in the
+# pending tree — same NotCommitted-residence rationale as MergeDialog.cs.
+Source/Parsek/MergeDialog.DialogText.cs
+
 # --- Phase 9 ERS-exempt files ---
 # SupersedeCommit.CommitTombstones scans the RAW ledger to BUILD the
 # tombstone set for the superseded subtree (design §6.6 step 4). Routing
